Added dev config to services.php and .gitignore
Following the same approach as the Invo tutorial.
A dev config file (app/config/config.dev.php) has been added to
the .gitignore. A check is done to see if this file is readable
when loading the configuration, if it is then this config is merged
with app/config/config.php.