Makes vokuro run on any subdirectory.
Since im testing  the vokuro project  i cannot test this on an sub-directory , i always have to re configure the document root on my webserver to point in out this on the folder where the vokuro project is ,... and this is very anoying.  

Note: you have to configure your .env file in this way APP_BASE_URI=/subfolder1/subfolder2/../subfolderN/YOURVOKURODIRECTORYNAME/ , look the path starts with slash (/) and ends with slash. 

This also works when the APP_BASE_URI its just one slash. please correct me if im re-inventing the wheel , because i cant no find any other fix about this issue.